+++ aeacus-src/structs.go
@@ -1,5 +1,14 @@
 package main
 
+import "github.com/urfave/cli"
+
+type metaConfig struct {
+	Cli     *cli.Context
+	TeamID  string
+	DirPath string
+	Config  scoringChecks
+}
+
 type imageData struct {
 	RunningTime int // change to time or smth idk
 	Score       int
